Associate an open menu group with its count. (#388)
* Associate an open menu group with its count.

Store open groups as an array of objects rather than an array of
strings. Each object has a count and group key. Prevents groups with
the same name from being treated as one, toggling all if one is toggled.

* Remove reducers.test.js file

This file is obsolete as reducers and their tests were moved to separate files

* Refactor code