Build menu categories on initialization and add to store (#491)
* Build menu categories on init and add to store

As part of the performance enhancement for gisida, the menu categories are now built during initialization to pervent the Menu component from re-building each time props change

* Remove unawnted properties for the menu categories

* Remove buildCategories export

* Solve bug cannot read id of undefined

* Handle html response when fetching layer

When fetching the layer object, if the layer obj is not found html is returned in production instead of 404 status. This scenario is handled in the changes

* Do not fetch layers again for other maps

Since other maps will be available when map-1 is loaded, we get the layer object from map-1 instead of making a HTTP GET

* Handle html fetch response in method loadJSON

* Add menu group ids after building categories

* When loading map-2 layers, only make visible default layers

* Update build

* Add layer to layerId lookup for ungrouped layers

* Add ungrouped layers to layerIds lookup only during map-1 initialization

* Update build

* Remove menu group count/id

The group count/id was used for making sure that menu groups with the same name in other categories do not open if a group is toggled open or do not close if a group is closed. Since the parent name is now attached to each group, we can make use of this instead

* Add build to .gitignore

* Update failing tests

* Adjust test case should create an action to toggle groups

* Move reducers test to src/store/reducers/tests, add tests for reducer CATEGORIES

* Add tests for action creator buildCategories

Also move action creator tests from test/store/actions to src/store/actions/tests

* Add tests for store/addMapLayerToStore

* Add tests for store/getMapLayer

* Add test cases for store/getMapLayer

* Refactor code

* Add doc strings